+++ src/logic/designer/connection.rs
@@ -1,10 +1,8 @@
 
 use std::sync::{Arc, Weak, RwLock};
 use super::super::error::LogicError;
-use super::super::descriptor::ConnectionDescriptor;
-use super::super::descriptor::TreatmentDescriptor;
+use super::super::descriptor::{ConnectionDescriptor, InputDescriptor, OutputDescriptor};
 use super::treatment::Treatment;
-use super::sequence::Sequence;
 
 #[derive(Debug)]
 pub enum IO {
@@ -32,8 +30,6 @@ impl PartialEq for IO {
 #[derive(Debug)]
 pub struct Connection {
 
-    sequence: Weak<RwLock<Sequence>>,
-
     descriptor: Arc<ConnectionDescriptor>,
 
     output_treatment: Option<IO>,
@@ -45,9 +41,8 @@ pub struct Connection {
 }
 
 impl Connection {
-    pub fn new(sequence: &Arc<RwLock<Sequence>>, descriptor: &Arc<ConnectionDescriptor>) -> Self {
+    pub fn new(descriptor: &Arc<ConnectionDescriptor>) -> Self {
         Self {
-            sequence: Arc::downgrade(sequence),
             descriptor: Arc::clone(descriptor),
             output_treatment: None,
             output_name: None,
@@ -60,141 +55,81 @@ impl Connection {
         &self.descriptor
     }
 
-    pub fn set_output(&mut self, treatment: &Arc<RwLock<Treatment>>, output: &Option<String>) -> Result<(), LogicError> {
+    /**
+     * Assign connection starting point.
+     * 
+     * Connections works as _treatment output_ -> _treatment input_
+     */
+    pub fn set_output(&mut self, treatment: &Arc<RwLock<Treatment>>, output: &OutputDescriptor) -> Result<(), LogicError> {
 
-        if output.is_none() {
-            if self.descriptor.output_type().is_none() {
-                self.output_treatment = Some(IO::Treatment(Arc::downgrade(treatment)));
-                self.output_name = None;
+        if output.datatype() == self.descriptor.output_type() {
 
-                Ok(())
-            }
-            else {
-                Err(LogicError::connection_output_required())
-            }
-        }
-        else if let Some(output_descriptor) = treatment.read().unwrap().descriptor().outputs().get(output.as_ref().unwrap()) {
+            self.output_treatment = Some(IO::Treatment(Arc::downgrade(treatment)));
+            self.output_name = Some(output.name().to_string());
 
-            if self.descriptor.output_type().is_none() {
-                Err(LogicError::connection_output_forbidden())
-            }
-            else if output_descriptor.datatype() == self.descriptor.output_type().as_ref().unwrap() {
-
-                self.output_treatment = Some(IO::Treatment(Arc::downgrade(treatment)));
-                self.output_name = output.as_ref().map(String::from);
-
-                Ok(())
-            }
-            else {
-                Err(LogicError::connection_output_unmatching_datatype())
-            }
+            Ok(())
         }
         else {
-            Err(LogicError::connection_output_not_found())
+            Err(LogicError::connection_output_unmatching_datatype())
         }
     }
 
-    pub fn set_self_output(&mut self, input_name: &Option<String>) -> Result<(), LogicError> {
+    /**
+     * Assign a self input as connection starting point (connection output).
+     * 
+     * `input` is seen from `Self`, that will be used as output for the connection
+     * (connections works as _treatment output_ -> _treatment input_).
+     */
+    pub fn set_self_output(&mut self, input: &InputDescriptor) -> Result<(), LogicError> {
         
-        if input_name.is_none() {
-            if self.descriptor.output_type().is_none() {
-                self.output_treatment = Some(IO::Sequence());
-                self.output_name = None;
+        if input.datatype() == self.descriptor.output_type() {
 
-                Ok(())
-            }
-            else {
-                Err(LogicError::connection_output_required())
-            }
-        }
-        else if let Some(input_descriptor) = self.sequence.upgrade().unwrap().read().unwrap()
-                                                .descriptor().inputs().get(input_name.as_ref().unwrap()) {
+            self.output_treatment = Some(IO::Sequence());
+            self.output_name = Some(input.name().to_string());
 
-            if self.descriptor.output_type().is_none() {
-                Err(LogicError::connection_output_forbidden())
-            }
-            else if input_descriptor.datatype() == self.descriptor.output_type().as_ref().unwrap() {
-
-                self.output_treatment = Some(IO::Sequence());
-                self.output_name = input_name.as_ref().map(String::from);
-
-                Ok(())
-            }
-            else {
-                Err(LogicError::connection_output_unmatching_datatype())
-            }
+            Ok(())
         }
         else {
-            Err(LogicError::connection_output_not_found())
+            Err(LogicError::connection_output_unmatching_datatype())
         }
     }
 
-    pub fn set_input(&mut self, treatment: &Arc<RwLock<Treatment>>, input: &Option<String>) -> Result<(), LogicError> {
+    /**
+     * Assign connection ending point.
+     * 
+     * Connections works as _treatment output_ -> _treatment input_
+     */
+    pub fn set_input(&mut self, treatment: &Arc<RwLock<Treatment>>, input: &InputDescriptor) -> Result<(), LogicError> {
 
-        if input.is_none() {
-            if self.descriptor.input_type().is_none() {
-                self.input_treatment = Some(IO::Treatment(Arc::downgrade(treatment)));
-                self.input_name = None;
+        if input.datatype() == self.descriptor.input_type() {
 
-                Ok(())
-            }
-            else {
-                Err(LogicError::connection_input_required())
-            }
-        }
-        else if let Some(input_descriptor) = treatment.read().unwrap().descriptor().inputs().get(input.as_ref().unwrap()) {
+            self.input_treatment = Some(IO::Treatment(Arc::downgrade(treatment)));
+            self.input_name = Some(input.name().to_string());
 
-            if self.descriptor.input_type().is_none() {
-                Err(LogicError::connection_input_forbidden())
-            }
-            else if input_descriptor.datatype() == self.descriptor.input_type().as_ref().unwrap() {
-
-                self.input_treatment = Some(IO::Treatment(Arc::downgrade(treatment)));
-                self.input_name = input.as_ref().map(String::from);
-
-                Ok(())
-            }
-            else {
-                Err(LogicError::connection_input_unmatching_datatype())
-            }
+            Ok(())
         }
         else {
-            Err(LogicError::connection_input_not_found())
+            Err(LogicError::connection_input_unmatching_datatype())
         }
     }
 
-    pub fn set_self_input(&mut self, output_name: &Option<String>) -> Result<(), LogicError> {
+    /**
+     * Assign a self ouput as connection ending point (connection input).
+     * 
+     * `output` is seen from `Self`, that will be used as input for the connection
+     * (connections works as _treatment output_ -> _treatment input_).
+     */
+    pub fn set_self_input(&mut self, output: &OutputDescriptor) -> Result<(), LogicError> {
 
-        if output_name.is_none() {
-            if self.descriptor.input_type().is_none() {
-                self.input_treatment = Some(IO::Sequence());
-                self.input_name = None;
+        if output.datatype() == self.descriptor.input_type() {
 
-                Ok(())
-            }
-            else {
-                Err(LogicError::connection_input_required())
-            }
-        }
-        else if let Some(output_descriptor) = self.sequence.upgrade().unwrap().read().unwrap()
-                                                    .descriptor().outputs().get(output_name.as_ref().unwrap()) {
+            self.input_treatment = Some(IO::Sequence());
+            self.input_name = Some(output.name().to_string());
 
-            if self.descriptor.input_type().is_none() {
-                Err(LogicError::connection_input_forbidden())
-            }
-            else if output_descriptor.datatype() == self.descriptor.input_type().as_ref().unwrap() {
-
-                self.input_treatment = Some(IO::Sequence());
-                self.input_name = output_name.as_ref().map(String::from);
-
-                Ok(())
-            }
-            else {
-                Err(LogicError::connection_input_unmatching_datatype())
-            }
+            Ok(())
         }
         else {
-            Err(LogicError::connection_input_not_found())
+            Err(LogicError::connection_input_unmatching_datatype())
         }
     }
 
@@ -224,28 +159,12 @@ impl Connection {
             return Err(LogicError::connection_input_not_set())
         }
 
-        // Check if descriptor require an output or not, then if one is assigned.
-        if let Some(_output) = self.descriptor.output_type() {
-            if self.output_name.is_none() {
-                return Err(LogicError::connection_output_required())
-            }
-        }
-        else {
-            if self.output_name.is_some() {
-                return Err(LogicError::connection_output_forbidden())
-            }
+        if self.output_name.is_none() {
+            return Err(LogicError::connection_output_required())
         }
 
-        // Check if descriptor require an input or not, then if one is assigned.
-        if let Some(_input) = self.descriptor.input_type() {
-            if self.input_name.is_none() {
-                return Err(LogicError::connection_input_required())
-            }
-        }
-        else {
-            if self.input_name.is_some() {
-                return Err(LogicError::connection_input_forbidden())
-            }
+        if self.input_name.is_none() {
+            return Err(LogicError::connection_input_required())
         }
 
         Ok(())