+++ src/logic/designer/sequence.rs
@@ -100,15 +100,15 @@ impl Sequence {
         }
     }
 
-    pub fn add_connection(&mut self, output_treatment: &str, output_name: &str, input_treatment: &str, input_name: &str) -> Result<Arc<RwLock<Connection>>, LogicError> {
+    pub fn add_connection(&mut self, output_treatment: &str, output_name: &str, input_treatment: &str, input_name: &str) -> Result<(), LogicError> {
         
         let rc_output_treatment;
-        let datatype_output;
+        let output;
         if let Some(pos_rc_output_treatment) = self.treatments.get(output_treatment) {
             rc_output_treatment = pos_rc_output_treatment;
 
             if let Some(pos_output) = rc_output_treatment.read().unwrap().descriptor().outputs().get(output_name) {
-                datatype_output = pos_output.datatype().clone();
+                output = pos_output.clone();
             }
             else {
                 return Err(LogicError::connection_output_not_found())
@@ -119,12 +119,12 @@ impl Sequence {
         }
 
         let rc_input_treatment;
-        let datatype_input;
+        let input;
         if let Some(pos_rc_input_treatment) = self.treatments.get(input_treatment) {
             rc_input_treatment = pos_rc_input_treatment;
 
             if let Some(pos_input) = rc_input_treatment.read().unwrap().descriptor().inputs().get(input_name) {
-                datatype_input = pos_input.datatype().clone();
+                input = pos_input.clone();
             }
             else {
                 return Err(LogicError::connection_input_not_found())
@@ -134,14 +134,17 @@ impl Sequence {
             return Err(LogicError::undeclared_treatment())
         }
 
-        if let Some(arc_connection_descriptor) = Connections::get(Some(datatype_output), Some(datatype_input)) {
+        if let Some(arc_connection_descriptor) = Connections::get(output.datatype(), input.datatype()) {
 
-            let connection = Connection::new(&self.auto_reference.upgrade().unwrap(), arc_connection_descriptor);
+            let mut connection = Connection::new(arc_connection_descriptor);
+
+            connection.set_input(rc_input_treatment, &input)?;
+            connection.set_output(rc_output_treatment, &output)?;
 
             let rc_connection = Arc::new(RwLock::new(connection));
-            self.connections.push(Arc::clone(&rc_connection));
+            self.connections.push(rc_connection);
 
-            return Ok(rc_connection)
+            return Ok(())
         }
         else {
             return Err(LogicError::unexisting_connexion_type())
@@ -173,32 +176,35 @@ impl Sequence {
         Ok(found)
     }
 
-    pub fn add_self_connection(&mut self, self_input_name: &str, self_output_name: &str) -> Result<Arc<RwLock<Connection>>, LogicError> {
+    pub fn add_self_connection(&mut self, self_input_name: &str, self_output_name: &str) -> Result<(), LogicError> {
 
-        let datatype_input_self;
+        let input_self;
         if let Some(pos_input) = self.descriptor.inputs().get(self_input_name) {
-            datatype_input_self = pos_input.datatype().clone();
+            input_self = pos_input.clone();
         }
         else {
             return Err(LogicError::connection_self_input_not_found())
         }
 
-        let datatype_output_self;
+        let output_self;
         if let Some(pos_output) = self.descriptor.outputs().get(self_output_name) {
-            datatype_output_self = pos_output.datatype().clone();
+            output_self = pos_output.clone();
         }
         else {
             return Err(LogicError::connection_self_output_not_found())
         }
 
-        if let Some(arc_connection_descriptor) = Connections::get(Some(datatype_input_self), Some(datatype_output_self)) {
+        if let Some(arc_connection_descriptor) = Connections::get(input_self.datatype(), output_self.datatype()) {
 
-            let connection = Connection::new(&self.auto_reference.upgrade().unwrap(), arc_connection_descriptor);
+            let mut connection = Connection::new(arc_connection_descriptor);
+
+            connection.set_self_output(&input_self)?;
+            connection.set_self_input(&output_self)?;
 
             let rc_connection = Arc::new(RwLock::new(connection));
-            self.connections.push(Arc::clone(&rc_connection));
+            self.connections.push(rc_connection);
 
-            return Ok(rc_connection)
+            return Ok(())
         }
         else {
             return Err(LogicError::unexisting_connexion_type())
@@ -230,23 +236,23 @@ impl Sequence {
         Ok(found)
     }
 
-    pub fn add_input_connection(&mut self, self_input_name: &str, input_treatment: &str, input_name: &str) -> Result<Arc<RwLock<Connection>>, LogicError> {
-        
-        let datatype_input_self;
+    pub fn add_input_connection(&mut self, self_input_name: &str, input_treatment: &str, input_name: &str) -> Result<(), LogicError> {
+
+        let input_self;
         if let Some(pos_input) = self.descriptor.inputs().get(self_input_name) {
-            datatype_input_self = pos_input.datatype().clone();
+            input_self = pos_input.clone();
         }
         else {
             return Err(LogicError::connection_self_input_not_found())
         }
 
         let rc_input_treatment;
-        let datatype_input_treatment;
+        let input;
         if let Some(pos_rc_input_treatment) = self.treatments.get(input_treatment) {
             rc_input_treatment = pos_rc_input_treatment;
 
             if let Some(pos_input) = rc_input_treatment.read().unwrap().descriptor().inputs().get(input_name) {
-                datatype_input_treatment = pos_input.datatype().clone();
+                input = pos_input.clone();
             }
             else {
                 return Err(LogicError::connection_input_not_found())
@@ -256,14 +262,17 @@ impl Sequence {
             return Err(LogicError::undeclared_treatment())
         }
 
-        if let Some(arc_connection_descriptor) = Connections::get(Some(datatype_input_self), Some(datatype_input_treatment)) {
+        if let Some(arc_connection_descriptor) = Connections::get(input_self.datatype(), input.datatype()) {
 
-            let connection = Connection::new(&self.auto_reference.upgrade().unwrap(), arc_connection_descriptor);
+            let mut connection = Connection::new(arc_connection_descriptor);
+
+            connection.set_self_output(&input_self)?;
+            connection.set_input(rc_input_treatment, &input)?;
 
             let rc_connection = Arc::new(RwLock::new(connection));
-            self.connections.push(Arc::clone(&rc_connection));
+            self.connections.push(rc_connection);
 
-            return Ok(rc_connection)
+            return Ok(())
         }
         else {
             return Err(LogicError::unexisting_connexion_type())
@@ -295,23 +304,23 @@ impl Sequence {
         Ok(found)
     }
 
-    pub fn add_output_connection(&mut self, self_output_name: &str, output_treatment: &str, output_name: &str) -> Result<Arc<RwLock<Connection>>, LogicError> {
-        
-        let datatype_output_self;
+    pub fn add_output_connection(&mut self, self_output_name: &str, output_treatment: &str, output_name: &str) -> Result<(), LogicError> {
+
+        let output_self;
         if let Some(pos_output) = self.descriptor.outputs().get(self_output_name) {
-            datatype_output_self = pos_output.datatype().clone();
+            output_self = pos_output.clone();
         }
         else {
             return Err(LogicError::connection_self_output_not_found())
         }
 
         let rc_output_treatment;
-        let datatype_output_treatment;
+        let output;
         if let Some(pos_rc_output_treatment) = self.treatments.get(output_treatment) {
             rc_output_treatment = pos_rc_output_treatment;
 
             if let Some(pos_output) = rc_output_treatment.read().unwrap().descriptor().outputs().get(output_name) {
-                datatype_output_treatment = pos_output.datatype().clone();
+                output = pos_output.clone();
             }
             else {
                 return Err(LogicError::connection_output_not_found())
@@ -321,14 +330,17 @@ impl Sequence {
             return Err(LogicError::undeclared_treatment())
         }
 
-        if let Some(arc_connection_descriptor) = Connections::get(Some(datatype_output_treatment), Some(datatype_output_self)) {
+        if let Some(arc_connection_descriptor) = Connections::get(output.datatype(), output_self.datatype()) {
 
-            let connection = Connection::new(&self.auto_reference.upgrade().unwrap(), arc_connection_descriptor);
+            let mut connection = Connection::new(arc_connection_descriptor);
+
+            connection.set_output(rc_output_treatment, &output)?;
+            connection.set_self_input(&output_self)?;
 
             let rc_connection = Arc::new(RwLock::new(connection));
-            self.connections.push(Arc::clone(&rc_connection));
+            self.connections.push(rc_connection);
 
-            return Ok(rc_connection)
+            return Ok(())
         }
         else {
             return Err(LogicError::unexisting_connexion_type())
@@ -385,6 +397,9 @@ impl Sequence {
         for connection in &self.connections {
 
             let borrowed_connection = connection.read().unwrap();
+
+            borrowed_connection.validate()?;
+
             match borrowed_connection.input_treatment().as_ref().unwrap() {
                 IO::Sequence() => {
                     *(outputs_satisfaction.get_mut(borrowed_connection.input_name().as_ref().unwrap()).unwrap()) += 1;