+++ src/script/semantic/sequence.rs
@@ -472,9 +472,7 @@ impl Sequence {
 
             let connection = rc_connection.read().unwrap();
 
-            let connection_designer = connection.get_designer(&mut rc_designer.write().unwrap())?;
-
-            connection.make_design(&connection_designer, &rc_designer)?;
+            connection.make_design(&mut rc_designer.write().unwrap())?;
         }
 
         wrap_logic_error!(